# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: StockWise API
    description: Учёт материальных средств и оборудования
    version: 1.0.0
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /equipments:
        get:
            tags:
                - equipments
            summary: list equipments
            description: Список оборудования с фильтрацией
            operationId: equipments#list
            parameters:
                - name: status
                  in: query
                  description: Фильтр по статусу
                  required: false
                  type: string
                  enum:
                    - exp
                    - exp_int
                    - exp_sp
                    - broken
                    - written_off
                - name: nomenclature_id
                  in: query
                  description: Фильтр по номенклатуре
                  required: false
                  type: integer
                  format: int32
                - name: location
                  in: query
                  description: Фильтр по местоположению
                  required: false
                  type: string
                - name: search
                  in: query
                  description: Поиск по наименованию/модели/серийнику
                  required: false
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/EquipmentList'
                        required:
                            - equipments
            schemes:
                - http
        post:
            tags:
                - equipments
            summary: create equipments
            description: Создать оборудование
            operationId: equipments#create
            parameters:
                - name: CreateRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/CreateEquipmentPayload'
                    required:
                        - inventory_number
                        - model_name
                        - status
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/Equipment'
                        required:
                            - inventory_number
                            - model_name
                            - status
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/EquipmentsCreateConflictResponseBody'
            schemes:
                - http
...
    /waybills:
        get:
            tags:
                - waybills
            summary: list waybills
            description: Список накладных
            operationId: waybills#list
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/WaybillList'
                        required:
                            - waybills
            schemes:
                - http
        post:
            tags:
                - waybills
            summary: create waybills
            description: Создать накладную (статус draft)
            operationId: waybills#create
            parameters:
                - name: CreateRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/CreateWaybillPayload'
                    required:
                        - number
                        - issue_date
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/Waybill'
                        required:
                            - number
                            - issue_date
                            - status
            schemes:
                - http
...
    /waybills/{id}/archive:
        post:
            tags:
                - waybills
            summary: archive waybills
            description: Архивировать накладную (signed → archived)
            operationId: waybills#archive
            parameters:
                - name: id
                  in: path
                  description: ID накладной
                  required: true
                  type: integer
                  format: int32
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Waybill'
                        required:
                            - number
                            - issue_date
                            - status
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/WaybillsArchiveNotFoundResponseBody'
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/WaybillsArchiveInvalidStatusResponseBody'
            schemes:
                - http
...
definitions:
...
    CreateEquipmentPayload:
        title: CreateEquipmentPayload
        type: object
        properties:
            arrival_date:
                type: string
                description: Дата поступления
                example: "2015-06-10"
                format: date
            form_number:
                type: string
                description: Номер формуляра
                example: Aut dolorum non aut et quod in.
            inventory_number:
                type: string
                description: Инвентарный номер
                example: Ad ex.
            location:
                type: string
                description: Место установки
                example: Accusamus quisquam aut voluptate voluptatem.
            manufacture_date:
                type: string
                description: Дата изготовления
                example: "1970-07-08"
                format: date
            model_name:
                type: string
                description: Наименование модели
                example: Consequatur corrupti quaerat fugiat.
            nomenclature_id:
                type: integer
                description: ID номенклатуры
                example: 1067063263
                format: int32
            notes:
                type: string
                description: Примечания
                example: Ut maxime et blanditiis.
            serial_number:
                type: string
                description: Заводской номер
                example: Odit atque ex.
            status:
                type: string
                description: Статус
                example: exp
                enum:
                    - exp
                    - exp_int
                    - exp_sp
                    - broken
                    - written_off
        example:
            arrival_date: "1993-12-28"
            form_number: Eos ratione odio.
            inventory_number: Architecto quae iusto saepe libero.
            location: Illum ea explicabo distinctio similique.
            manufacture_date: "2006-08-07"
            model_name: Enim neque eius molestiae veniam deleniti impedit.
            nomenclature_id: 761832175
            notes: Non non sint explicabo ut.
            serial_number: Placeat placeat dolorum autem.
            status: exp_sp
        required:
            - inventory_number
            - model_name
            - status
    CreateWaybillPayload:
        title: CreateWaybillPayload
        type: object
        properties:
            from_dept:
                type: integer
                description: Код подразделения-отправителя
                example: 1968344135
                format: int32
            issue_date:
                type: string
                description: Дата документа
                example: "1981-12-28"
                format: date
            items:
                type: array
                items:
                    $ref: '#/definitions/WaybillsEquipment'
                description: Позиции накладной
                example:
                    - equipment_id: 2016065243
                      waybill_id: 664585174
                    - equipment_id: 2016065243
                      waybill_id: 664585174
                    - equipment_id: 2016065243
                      waybill_id: 664585174
            number:
                type: string
                description: Номер документа
                example: Rerum dolor magnam.
            to_dept:
                type: integer
                description: Код подразделения-получателя
                example: 1439361675
                format: int32
        example:
            from_dept: 1630865970
            issue_date: "2012-05-04"
            items:
                - equipment_id: 2016065243
                  waybill_id: 664585174
                - equipment_id: 2016065243
                  waybill_id: 664585174
            number: Neque maxime ut.
            to_dept: 612290589
        required:
            - number
            - issue_date
...
    Equipment:
        title: Equipment
        type: object
        properties:
            arrival_date:
                type: string
                description: Дата поступления
                example: "2012-03-05"
                format: date
            assignment:
                $ref: '#/definitions/AssignmentInfo'
            form_number:
                type: string
                description: Номер формуляра
                example: Et ut ea culpa.
            inventory_number:
                type: string
                description: Инвентарный номер
                example: ИТ00205
                pattern: ^ИТ\d{5}$
            location:
                type: string
                description: Место установки
                example: "219"
            manufacture_date:
                type: string
                description: Дата изготовления
                example: "1996-02-15"
                format: date
            model_name:
                type: string
                description: Наименование модели
                example: Ноутбук Clevo 15"
            nomenclature:
                $ref: '#/definitions/Nomenclature'
            notes:
                type: string
                description: Примечания
                example: Ipsam repellendus ea expedita autem non temporibus.
            serial_number:
                type: string
                description: Заводской номер
                example: Aut distinctio autem qui nostrum.
            status:
                type: string
                description: Статус
                example: written_off
                enum:
                    - exp
                    - exp_int
                    - exp_sp
                    - broken
                    - written_off
        description: Оборудование
        example:
            arrival_date: "1990-04-15"
            assignment:
                card_number: 195554449
                dept_name: Qui quo voluptatem dicta.
                full_name: Багликов А.С.
                operator_comment: Ut minus omnis aut suscipit.
                target_type: warehouse
            form_number: Deleniti aliquid debitis esse.
            inventory_number: ИТ00205
            location: "219"
            manufacture_date: "1977-06-20"
            model_name: Ноутбук Clevo 15"
            nomenclature:
                code: 08.01.02.00.00/0041
                name: Ноутбук Clevo 15"
            notes: A et tenetur totam sequi.
            serial_number: Ad sit molestias.
            status: broken
        required:
            - inventory_number
            - model_name
            - status
    EquipmentList:
        title: EquipmentList
        type: object
        properties:
            equipments:
                type: array
                items:
                    $ref: '#/definitions/Equipment'
                example:
                    - arrival_date: "2002-02-27"
                      assignment:
                        card_number: 195554449
                        dept_name: Qui quo voluptatem dicta.
                        full_name: Багликов А.С.
                        operator_comment: Ut minus omnis aut suscipit.
                        target_type: warehouse
                      form_number: Quis iste velit veritatis numquam soluta.
                      inventory_number: ИТ00205
                      location: "219"
                      manufacture_date: "1988-05-21"
                      model_name: Ноутбук Clevo 15"
                      nomenclature:
                        code: 08.01.02.00.00/0041
                        name: Ноутбук Clevo 15"
                      notes: Saepe magnam facilis assumenda eos voluptatem voluptatem.
                      serial_number: Culpa facere voluptas explicabo quis.
                      status: exp
                    - arrival_date: "2002-02-27"
                      assignment:
                        card_number: 195554449
                        dept_name: Qui quo voluptatem dicta.
                        full_name: Багликов А.С.
                        operator_comment: Ut minus omnis aut suscipit.
                        target_type: warehouse
                      form_number: Quis iste velit veritatis numquam soluta.
                      inventory_number: ИТ00205
                      location: "219"
                      manufacture_date: "1988-05-21"
                      model_name: Ноутбук Clevo 15"
                      nomenclature:
                        code: 08.01.02.00.00/0041
                        name: Ноутбук Clevo 15"
                      notes: Saepe magnam facilis assumenda eos voluptatem voluptatem.
                      serial_number: Culpa facere voluptas explicabo quis.
                      status: exp
        example:
            equipments:
                - arrival_date: "2002-02-27"
                  assignment:
                    card_number: 195554449
                    dept_name: Qui quo voluptatem dicta.
                    full_name: Багликов А.С.
                    operator_comment: Ut minus omnis aut suscipit.
                    target_type: warehouse
                  form_number: Quis iste velit veritatis numquam soluta.
                  inventory_number: ИТ00205
                  location: "219"
                  manufacture_date: "1988-05-21"
                  model_name: Ноутбук Clevo 15"
                  nomenclature:
                    code: 08.01.02.00.00/0041
                    name: Ноутбук Clevo 15"
                  notes: Saepe magnam facilis assumenda eos voluptatem voluptatem.
                  serial_number: Culpa facere voluptas explicabo quis.
                  status: exp
                - arrival_date: "2002-02-27"
                  assignment:
                    card_number: 195554449
                    dept_name: Qui quo voluptatem dicta.
                    full_name: Багликов А.С.
                    operator_comment: Ut minus omnis aut suscipit.
                    target_type: warehouse
                  form_number: Quis iste velit veritatis numquam soluta.
                  inventory_number: ИТ00205
                  location: "219"
                  manufacture_date: "1988-05-21"
                  model_name: Ноутбук Clevo 15"
                  nomenclature:
                    code: 08.01.02.00.00/0041
                    name: Ноутбук Clevo 15"
                  notes: Saepe magnam facilis assumenda eos voluptatem voluptatem.
                  serial_number: Culpa facere voluptas explicabo quis.
                  status: exp
                - arrival_date: "2002-02-27"
                  assignment:
                    card_number: 195554449
                    dept_name: Qui quo voluptatem dicta.
                    full_name: Багликов А.С.
                    operator_comment: Ut minus omnis aut suscipit.
                    target_type: warehouse
                  form_number: Quis iste velit veritatis numquam soluta.
                  inventory_number: ИТ00205
                  location: "219"
                  manufacture_date: "1988-05-21"
                  model_name: Ноутбук Clevo 15"
                  nomenclature:
                    code: 08.01.02.00.00/0041
                    name: Ноутбук Clevo 15"
                  notes: Saepe magnam facilis assumenda eos voluptatem voluptatem.
                  serial_number: Culpa facere voluptas explicabo quis.
                  status: exp
                - arrival_date: "2002-02-27"
                  assignment:
                    card_number: 195554449
                    dept_name: Qui quo voluptatem dicta.
                    full_name: Багликов А.С.
                    operator_comment: Ut minus omnis aut suscipit.
                    target_type: warehouse
                  form_number: Quis iste velit veritatis numquam soluta.
                  inventory_number: ИТ00205
                  location: "219"
                  manufacture_date: "1988-05-21"
                  model_name: Ноутбук Clevo 15"
                  nomenclature:
                    code: 08.01.02.00.00/0041
                    name: Ноутбук Clevo 15"
                  notes: Saepe magnam facilis assumenda eos voluptatem voluptatem.
                  serial_number: Culpa facere voluptas explicabo quis.
                  status: exp
        required:
            - equipments
    EquipmentsCreateConflictResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: true
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: false
        description: create_conflict_response_body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: false
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
...
    Waybill:
        title: Waybill
        type: object
        properties:
            from_dept:
                type: integer
                description: Код подразделения-отправителя
                example: 1904907712
                format: int32
            issue_date:
                type: string
                description: Дата документа
                example: "1990-09-11"
                format: date
            items:
                type: array
                items:
                    $ref: '#/definitions/WaybillsEquipment'
                description: Позиции накладной
                example:
                    - equipment_id: 1232401099
                      waybill_id: 356114870
                    - equipment_id: 1232401099
                      waybill_id: 356114870
                    - equipment_id: 1232401099
                      waybill_id: 356114870
                    - equipment_id: 1232401099
                      waybill_id: 356114870
            number:
                type: string
                description: Номер документа
                example: Накладная №123
            status:
                type: string
                description: Статус
                example: signed
                enum:
                    - draft
                    - signed
                    - archived
            to_dept:
                type: integer
                description: Код подразделения-получателя
                example: 544755368
                format: int32
        description: Накладная
        example:
            from_dept: 729853037
            issue_date: "2009-02-16"
            items:
                - equipment_id: 1232401099
                  waybill_id: 356114870
                - equipment_id: 1232401099
                  waybill_id: 356114870
                - equipment_id: 1232401099
                  waybill_id: 356114870
                - equipment_id: 1232401099
                  waybill_id: 356114870
            number: Накладная №123
            status: archived
            to_dept: 537770379
        required:
            - number
            - issue_date
            - status
    WaybillList:
        title: WaybillList
        type: object
        properties:
            waybills:
                type: array
                items:
                    $ref: '#/definitions/Waybill'
                example:
                    - from_dept: 1203015280
                      issue_date: "1987-12-11"
                      items:
                        - equipment_id: 1232401099
                          waybill_id: 356114870
                        - equipment_id: 1232401099
                          waybill_id: 356114870
                      number: Накладная №123
                      status: archived
                      to_dept: 687976267
                    - from_dept: 1203015280
                      issue_date: "1987-12-11"
                      items:
                        - equipment_id: 1232401099
                          waybill_id: 356114870
                        - equipment_id: 1232401099
                          waybill_id: 356114870
                      number: Накладная №123
                      status: archived
                      to_dept: 687976267
                    - from_dept: 1203015280
                      issue_date: "1987-12-11"
                      items:
                        - equipment_id: 1232401099
                          waybill_id: 356114870
                        - equipment_id: 1232401099
                          waybill_id: 356114870
                      number: Накладная №123
                      status: archived
                      to_dept: 687976267
                    - from_dept: 1203015280
                      issue_date: "1987-12-11"
                      items:
                        - equipment_id: 1232401099
                          waybill_id: 356114870
                        - equipment_id: 1232401099
                          waybill_id: 356114870
                      number: Накладная №123
                      status: archived
                      to_dept: 687976267
        example:
            waybills:
                - from_dept: 1203015280
                  issue_date: "1987-12-11"
                  items:
                    - equipment_id: 1232401099
                      waybill_id: 356114870
                    - equipment_id: 1232401099
                      waybill_id: 356114870
                  number: Накладная №123
                  status: archived
                  to_dept: 687976267
                - from_dept: 1203015280
                  issue_date: "1987-12-11"
                  items:
                    - equipment_id: 1232401099
                      waybill_id: 356114870
                    - equipment_id: 1232401099
                      waybill_id: 356114870
                  number: Накладная №123
                  status: archived
                  to_dept: 687976267
                - from_dept: 1203015280
                  issue_date: "1987-12-11"
                  items:
                    - equipment_id: 1232401099
                      waybill_id: 356114870
                    - equipment_id: 1232401099
                      waybill_id: 356114870
                  number: Накладная №123
                  status: archived
                  to_dept: 687976267
                - from_dept: 1203015280
                  issue_date: "1987-12-11"
                  items:
                    - equipment_id: 1232401099
                      waybill_id: 356114870
                    - equipment_id: 1232401099
                      waybill_id: 356114870
                  number: Накладная №123
                  status: archived
                  to_dept: 687976267
        required:
            - waybills
...